Factor Cursor methods into appropriate subclass.
