+++ src/stream/Cursor.cpp
@@ -1,27 +1,20 @@
-/* -*- C++ -*- */
-/*
- * Copyright 2016 WebAssembly Community Group participants
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// -*- C++ -*-
+//
+// Copyright 2016 WebAssembly Community Group participants
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #include "stream/Cursor.h"
-
-namespace {
-#ifndef NDEBUG
-static constexpr uint32_t MaxBitsAllowed = sizeof(uint32_t) * CHAR_BIT;
-#endif
-}
 
 namespace wasm {
 
@@ -61,45 +54,6 @@ void Cursor::writeFillBuffer() {
   size_t BufferSize = Que->writeToPage(CurAddress, Page::Size, *this);
   if (BufferSize == 0)
     fatal("Write failed!\n");
-}
-
-uint32_t Cursor::readBits(uint32_t NumBits) {
-  assert(NumBits <= MaxBitsAllowed);
-  uint32_t Value = 0;
-  while (NumBits != 0) {
-    BitsInByteType AvailBits = CurByte.getReadBitsRemaining();
-    if (NumBits <= AvailBits) {
-      Value = (Value << NumBits) | CurByte.readBits(NumBits);
-      return Value;
-    }
-    if (!CurByte.isEmpty()) {
-      Value = (Value << AvailBits) | CurByte.getValue();
-      NumBits -= AvailBits;
-    }
-    CurByte.setByte(readByte());
-  }
-  return Value;
-}
-
-void Cursor::writeBits(uint32_t Value, uint32_t NumBits) {
-  assert(NumBits <= MaxBitsAllowed);
-  while (NumBits > 0) {
-    const BitsInByteType AvailBits = CurByte.getWriteBitsRemaining();
-    if (AvailBits >= NumBits) {
-      CurByte.writeBits(Value, NumBits);
-      if (AvailBits == NumBits) {
-        writeByte(CurByte.getValue());
-        CurByte.reset();
-      }
-      return;
-    }
-    uint32_t Shift = NumBits - AvailBits;
-    CurByte.writeBits(Value >> Shift, AvailBits);
-    Value &= (uint32_t(1) << Shift) - 1;
-    NumBits -= AvailBits;
-    writeByte(CurByte.getValue());
-    CurByte.reset();
-  }
 }
 
 FILE* Cursor::describe(FILE* File, bool IncludeDetail) {
